shared/python/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# Kafka
aiokafka>=0.10.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
authlib>=1.3.0
httpx>=0.26.0

# AWS
boto3>=1.34.0

# QR Code generation
qrcode[pil]>=7.4.0
Pillow>=10.2.0

# User agent parsing
user-agents>=2.2.0

# GeoIP
geoip2>=4.8.0

# ClickHouse
clickhouse-driver>=0.2.6

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
